validation/requirements.txt
pydantic>=2.6.0
requests>=2.31.0
scikit-learn>=1.4.0
joblib>=1.3.0
numpy>=1.26.0
minio
python-dotenv==1.0.1